ml_projects/Fake-Driving-License/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
imageio==2.21.2
imutils==0.5.4
networkx==2.8.6
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyWavelets==1.3.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
tifffile==2022.8.12
urllib3==1.26.12
